+++ core/src/main/java/nl/nn/adapterframework/configuration/IbisManager.java
@@ -59,7 +59,7 @@ public interface IbisManager {
      * @param receiverName
      * @param commandIssuedBy
      */
-    void handleAdapter(String action, String adapterName, String receiverName, String commandIssuedBy);
+    void handleAdapter(String action, String configurationName, String adapterName, String receiverName, String commandIssuedBy);
     /**
      * Start an already configured Configuration
      *