+++ RELEASES.md
@@ -37,6 +37,7 @@ Upcoming
     - Cache original and loaded configuration
     - Add getConfiguration to Adapter
     - Add getIbisManager to Configuration
+- Make it possible to stop and start all adapters per configuration